engine-eds: fix retrieving custom alarm sound path
ECal 2.0 returns the list of attachments as a GSList (a singly linked
list). I'm not sure why, but the logic for iterating the list is
completely incorrect. Fixing that fixes custom alarm sound.

A test is added to catch this case.

Bug-UBports: https://gitlab.com/ubports/development/apps/lomiri-clock-app/-/issues/183